+++ src/v2/templates/cart/cartfunctions.php
@@ -1,0 +1,73 @@
+<!-- BEGIN CART FUNCTIONS -->
+<?php
+
+use ChurchCRM\dto\Cart;
+use ChurchCRM\SessionUser;
+use ChurchCRM\dto\SystemConfig;
+use ChurchCRM\dto\SystemURLs;
+?>
+<div class="box">
+  <div class="box-header with-border">
+    <h3 class="box-title">Cart Functions</h3>
+  </div>
+  <div class="box-body">
+    <a href="#" id="emptyCart" class="btn btn-app emptyCart"><i class="fa fa-trash"></i><?= gettext('Empty Cart') ?></a>
+    <?php if ($_SESSION['user']->isManageGroupsEnabled()) {
+      ?>
+      <a id="emptyCartToGroup" class="btn btn-app"><i class="fa fa-object-ungroup"></i><?= gettext('Empty Cart to Group') ?></a>
+      <?php
+    }
+    if ($_SESSION['user']->isAddRecordsEnabled()) {
+      ?>
+      <a href="CartToFamily.php" class="btn btn-app"><i
+          class="fa fa-users"></i><?= gettext('Empty Cart to Family') ?></a>
+        <?php }
+      ?>
+    <a href="CartToEvent.php" class="btn btn-app"><i
+        class="fa fa-ticket"></i><?= gettext('Empty Cart to Event') ?></a>
+
+    <?php if (SessionUser::getUser()->isCSVExport()) {
+      ?>
+      <a href="CSVExport.php?Source=cart" class="btn btn-app"><i
+          class="fa fa-file-excel-o"></i><?= gettext('CSV Export') ?></a>
+        <?php }
+      ?>
+    <a href="MapUsingGoogle.php?GroupID=0" class="btn btn-app"><i
+        class="fa fa-map-marker"></i><?= gettext('Map Cart') ?></a>
+    <a href="Reports/NameTags.php?labeltype=74536&labelfont=times&labelfontsize=36" class="btn btn-app"><i
+        class="fa fa-file-pdf-o"></i><?= gettext('Name Tags') ?></a>
+      <?php
+     
+          if ($bEmailMailto) { // Does user have permission to email groups
+            // Display link
+            echo "<a href='mailto:" . mb_substr($sEmailLink, 0, -3) . "' class='btn btn-app'><i class='fa fa-send-o'></i>" . gettext('Email Cart') . '</a>';
+            echo "<a href='mailto:?bcc=" . mb_substr($sEmailLink, 0, -3) . "' class='btn btn-app'><i class='fa fa-send'></i>" . gettext('Email (BCC)') . '</a>';
+          }
+        
+        if ($sPhoneLink) {
+          if ($bEmailMailto) { // Does user have permission to email groups
+            // Display link
+            echo '<a href="javascript:void(0)" onclick="allPhonesCommaD()" class="btn btn-app"><i class="fa fa-mobile-phone"></i>' . gettext("Text Cart");
+            echo '<script nonce="' . SystemURLs::getCSPNonce() . '">function allPhonesCommaD() {prompt("Press CTRL + C to copy all group members\' phone numbers", "' . mb_substr($sPhoneLink, 0, -2) . '")};</script>';
+          }
+        }
+        ?>
+      <a href="DirectoryReports.php?cartdir=Cart+Directory" class="btn btn-app"><i
+          class="fa fa-book"></i><?= gettext('Create Directory From Cart') ?></a>
+
+      <script nonce="<?= SystemURLs::getCSPNonce() ?>" ><!--
+                  function codename() {
+        if (document.labelform.bulkmailpresort.checked) {
+          document.labelform.bulkmailquiet.disabled = false;
+        } else {
+          document.labelform.bulkmailquiet.disabled = true;
+          document.labelform.bulkmailquiet.checked = false;
+        }
+      }
+
+      //-->
+                </SCRIPT>
+              </div>
+              <!-- /.box-body -->
+            </div>
+    <!-- END CART FUNCTIONS -->
