+++ tests/behat/features/Cart/ViewCart.feature
@@ -5,7 +5,7 @@ Feature: Cart
 
   Scenario: Cart Add and Remove
     Given I am authenticated as "admin" using "changeme"
-    And I am on "CartView.php"
+    And I am on "/v2/cart"
     Then I should see "You have no items in your cart"
     And I am on "PersonView.php?PersonID=1"
     And I click "#AddPersonToCart"
